convert into a bug report template specifically
Using the check boxes for identifying that it's a bug report or a feature request causes a progress checklist to show up as if the boxes signified a to-do list. Changing that selection to be this way lets it still get the same desired effect, while not getting that unfortunate side effect of a phantom to-do list.